requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alpaca-py==0.39.4
alpaca-trade-api==3.2.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.3.0
binance==0.3.5
blinker==1.9.0
Bottleneck==1.6.0
cachetools==6.2.2
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==46.0.3
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.17.0
email_validator==2.2.0
fastapi==0.115.11
fastapi-cli==0.0.7
filetype==1.2.0
Flask==3.1.0
frozenlist==1.5.0
git-filter-repo==2.47.0
google-ai-generativelanguage==0.9.0
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-aiplatform==1.127.0
google-cloud-bigquery==3.38.0
google-cloud-core==2.5.0
google-cloud-resource-manager==1.15.0
google-cloud-storage==3.5.0
google-crc32c==1.7.1
google-genai==1.50.1
google-resumable-media==2.7.2
googleapis-common-protos==1.72.0
groq==0.34.1
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.3
idna==3.10
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==1.0.5
langchain-brightdata==0.1.3
langchain-core==1.0.4
langchain-google-genai==3.0.3
langchain-google-vertexai==3.0.3
langchain-mcp-adapters==0.1.13
langchain-openai==1.0.2
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.4
langgraph-sdk==0.2.9
langgraph-supervisor==0.0.30
langsmith==0.4.42
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.21.1
mdurl==0.1.2
msgpack==1.0.3
multidict==6.4.3
numexpr==2.14.1
numpy==2.2.4
openai==2.8.0
orjson==3.10.15
ormsgpack==1.12.0
packaging==24.2
pandas==2.2.3
pluggy==1.6.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.33.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic-extra-types==2.10.3
pydantic-settings==2.8.1
pydantic_core==2.41.5
Pygments==2.19.1
PyJWT==2.10.1
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.2
pywin32==311
PyYAML==6.0.1
referencing==0.37.0
regex==2025.11.3
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.28.0
rsa==4.9.1
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sse-starlette==3.0.3
sseclient-py==1.8.0
starlette==0.46.1
ta==0.11.0
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typer==0.15.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.10.0
urllib3==1.26.20
uvicorn==0.34.0
validators==0.35.0
watchfiles==1.0.4
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
xxhash==3.6.0
yarl==1.20.0
zstandard==0.25.0
